Python implementation of haproxy_status.sh
We seem to be running into issues where haproxy_status.sh is refusing
connections, resulting in failed connections to the primary due to
haproxy thinking all backends are unhealthy. This replaces the shell
script/netcat business with a Python script that does the same
thing. Hopefully it will be more reliable.